move GetEntityByID test to its own file
